Rename file to createControlPanel
